Add test cases for blob file GC with online compaction (WIP)
This commit refactors the compaction test code by splitting the existing compaction_test.cpp file into two separate files:

  - compaction_test_fixture.h: Contains the test fixture class, helper functions, and common includes used by the compaction tests.
  - compaction_test.cpp: Contains only the TEST_F cases that reference the fixture defined in the header.
