+++ src/limestone/dblog_scan.h
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#pragma once
 
 #include <boost/filesystem.hpp>
