+++ test/test_root.h
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#pragma once
 
 #include <gtest/gtest.h>
 #include <glog/logging.h>
